requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
minestat==2.6.1
Pillow==10.0.0
python-dotenv==1.0.0
Requests==2.31.0
lxml